+++ utopia-remix/app/routes-test/internal.projects.$id.github.branches.$owner.$repo.branches.$branch.spec.ts
@@ -16,7 +16,7 @@ import path from 'path'
 import urlJoin from 'url-join'
 import * as fs from 'fs'
 import type { ApiSuccess } from '../types'
-import type { BranchResponse } from '../util/github-branch-contents'
+import type { BranchResponse } from '../util/github-branch-contents.server'
 
 describe('get branch project contents', () => {
   let mockOctokit: jest.SpyInstance
