+++ utopia-remix/app/routes/internal.projects.$id.github.branches.$owner.$repo.branch.$branch.tsx
@@ -6,7 +6,7 @@ import type { ExistingAsset } from '../types'
 import { UserProjectPermission } from '../types'
 import { Status } from '../util/statusCodes'
 import { getGithubAuthentication } from '../models/githubAuthentication.server'
-import { getBranchProjectContents } from '../util/github-branch-contents'
+import { getBranchProjectContents } from '../util/github-branch-contents.server'
 import { newOctokitClient, wrapGithubAPIRequest } from '../util/github'
 
 export function loader(args: LoaderFunctionArgs) {
